+++ ImageAnalyser.py
